critical bug fixing : tank now is able to process input-Move correctly (TODO: check if Fire is well handled now)
