+++ robots-module/src/test/java/orwell/tank/SplitTest.java
@@ -7,6 +7,7 @@ import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import orwell.tank.utils.Split;
 
 import java.util.List;
 
@@ -16,8 +17,8 @@ import static org.junit.Assert.*;
  * Created by Michaël Ludmann on 6/11/15.
  */
 @RunWith(JUnit4.class)
-public class UtilsTest {
-    private final static Logger logback = LoggerFactory.getLogger(UtilsTest.class);
+public class SplitTest {
+    private final static Logger logback = LoggerFactory.getLogger(SplitTest.class);
     private static final char SEPARATOR = ' ';
 
     @Before
@@ -28,7 +29,7 @@ public class UtilsTest {
     @Test
     public void testSplit_cSEPc_limit3() throws Exception {
         final String string = "x y";
-        final List<String> list = Utils.split(SEPARATOR, string, 3);
+        final List<String> list = Split.split(SEPARATOR, string, 3);
         assertEquals(2, list.size());
         assertEquals("x", list.get(0));
         assertEquals("y", list.get(1));
@@ -37,7 +38,7 @@ public class UtilsTest {
     @Test
     public void testSplit_cSEPc_limit2() throws Exception {
         final String string = "x y";
-        final List<String> list = Utils.split(SEPARATOR, string, 2);
+        final List<String> list = Split.split(SEPARATOR, string, 2);
         assertEquals(2, list.size());
         assertEquals("x", list.get(0));
         assertEquals("y", list.get(1));
@@ -46,7 +47,7 @@ public class UtilsTest {
     @Test
     public void testSplit_cSEPc_limit1() throws Exception {
         final String string = "x y";
-        final List<String> list = Utils.split(SEPARATOR, string, 1);
+        final List<String> list = Split.split(SEPARATOR, string, 1);
         assertEquals(1, list.size());
         assertEquals("x", list.get(0));
     }
@@ -54,14 +55,14 @@ public class UtilsTest {
     @Test
     public void testSplit_cSEPc_limit0() throws Exception {
         final String string = "x y";
-        final List<String> list = Utils.split(SEPARATOR, string, 0);
+        final List<String> list = Split.split(SEPARATOR, string, 0);
         assertEquals(0, list.size());
     }
 
     @Test
     public void testSplit_SEPc_limit2() throws Exception {
         final String string = " y";
-        final List<String> list = Utils.split(SEPARATOR, string, 2);
+        final List<String> list = Split.split(SEPARATOR, string, 2);
         assertEquals(2, list.size());
         assertEquals("", list.get(0));
         assertEquals("y", list.get(1));
@@ -70,7 +71,7 @@ public class UtilsTest {
     @Test
     public void testSplit_cSEP_limit2() throws Exception {
         final String string = "x ";
-        final List<String> list = Utils.split(SEPARATOR, string, 2);
+        final List<String> list = Split.split(SEPARATOR, string, 2);
         assertEquals(2, list.size());
         assertEquals("x", list.get(0));
         assertEquals("", list.get(1));
@@ -79,7 +80,7 @@ public class UtilsTest {
     @Test
     public void testSplit_SEP_limit2() throws Exception {
         final String string = " ";
-        final List<String> list = Utils.split(SEPARATOR, string, 2);
+        final List<String> list = Split.split(SEPARATOR, string, 2);
         assertEquals(2, list.size());
         assertEquals("", list.get(0));
         assertEquals("", list.get(1));
@@ -88,7 +89,7 @@ public class UtilsTest {
     @Test
     public void testSplit_cSEPSEP_limit3() throws Exception {
         final String string = "x  ";
-        final List<String> list = Utils.split(SEPARATOR, string, 3);
+        final List<String> list = Split.split(SEPARATOR, string, 3);
         assertEquals(3, list.size());
         assertEquals("x", list.get(0));
         assertEquals("", list.get(1));
@@ -98,7 +99,7 @@ public class UtilsTest {
     @Test
     public void testSplit_SEPcSEP_limit3() throws Exception {
         final String string = " x ";
-        final List<String> list = Utils.split(SEPARATOR, string, 3);
+        final List<String> list = Split.split(SEPARATOR, string, 3);
         assertEquals(3, list.size());
         assertEquals("", list.get(0));
         assertEquals("x", list.get(1));
@@ -108,7 +109,7 @@ public class UtilsTest {
     @Test
     public void testSplit_SEPSEP_limit3() throws Exception {
         final String string = "  ";
-        final List<String> list = Utils.split(SEPARATOR, string, 3);
+        final List<String> list = Split.split(SEPARATOR, string, 3);
         assertEquals(3, list.size());
         assertEquals("", list.get(0));
         assertEquals("", list.get(1));
@@ -118,7 +119,7 @@ public class UtilsTest {
     @Test
     public void testSplit_cSEPSEPc_limit3() throws Exception {
         final String string = "x  y";
-        final List<String> list = Utils.split(SEPARATOR, string, 3);
+        final List<String> list = Split.split(SEPARATOR, string, 3);
         assertEquals(3, list.size());
         assertEquals("x", list.get(0));
         assertEquals("", list.get(1));
@@ -128,7 +129,7 @@ public class UtilsTest {
     @Test
     public void testSplit_cc_limit2() throws Exception {
         final String string = "xy";
-        final List<String> list = Utils.split(SEPARATOR, string, 2);
+        final List<String> list = Split.split(SEPARATOR, string, 2);
         assertEquals(1, list.size());
         assertEquals(string, list.get(0));
     }
@@ -136,7 +137,7 @@ public class UtilsTest {
     @Test
     public void testSplit_ccSEPcccSEPcc_limit3() throws Exception {
         final String string = "xy abc de";
-        final List<String> list = Utils.split(SEPARATOR, string, 3);
+        final List<String> list = Split.split(SEPARATOR, string, 3);
         assertEquals(3, list.size());
         assertEquals("xy", list.get(0));
         assertEquals("abc", list.get(1));
